added authorization

+++ src/tasks/tasks.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  InternalServerErrorException,
+} from '@nestjs/common';
 
 import { TaskStatus } from './task-status.enum';
 import { CreateTaskDTO } from './dto/create-task.dto';
@@ -6,6 +10,7 @@ import { GetTasksFilterDTO } from './dto/get-tasks-filter.dto';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,37 +19,50 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(getTasksFilterDTO: GetTasksFilterDTO): Promise<Task[]> {
-    return await this.taskRepository.getTasks(getTasksFilterDTO);
+  async getTasks(
+    getTasksFilterDTO: GetTasksFilterDTO,
+    user: User,
+  ): Promise<Task[]> {
+    return await this.taskRepository.getTasks(getTasksFilterDTO, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const foundTask = await this.taskRepository.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const foundTask = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!foundTask) {
       throw new NotFoundException(`Task with id "${id}" not found`);
     }
     return foundTask;
   }
 
-  createTask(createTaskDTO: CreateTaskDTO): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDTO);
+  createTask(createTaskDTO: CreateTaskDTO, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDTO, user);
   }
 
-  async updateTaskStatus(id: number, newStatus: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(
+    id: number,
+    newStatus: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = newStatus;
     await task.save();
     return task;
   }
-  async deleteTaskById(id: number): Promise<void> {
+  async deleteTaskById(id: number, user: User): Promise<void> {
     // requires 2 calls to the db
     // const taskFound = await this.getTaskById(id);
     // return await this.taskRepository.remove(taskFound);
 
     // requires only 1 call to the db
-    const result = await this.taskRepository.delete(id);
-    if (result.affected === 0) {
-      throw new NotFoundException(`Task with id "${id}" not found`);
+    try {
+      const result = await this.taskRepository.delete({ id, userId: user.id });
+      if (result.affected === 0) {
+        throw new NotFoundException(`Task with id "${id}" not found`);
+      }
+    } catch (error) {
+      console.log(error.message);
     }
   }
 }
